Restructure BindingInterfacesWithInnerDependencies domain logic test
