+++ tests/Integration/Framework/BindingInterfacesWithInnerDependencies/LocalConfig/Domain/Greeter/CorrectCompanyGenerator.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GacelaTest\Integration\Framework\BindingInterfacesWithInnerDependencies\LocalConfig\Infrastructure;
+namespace GacelaTest\Integration\Framework\BindingInterfacesWithInnerDependencies\LocalConfig\Domain\Greeter;
 
 use GacelaTest\Integration\Framework\BindingInterfacesWithInnerDependencies\LocalConfig\Domain\GreeterGeneratorInterface;
 
@@ -10,6 +10,10 @@ final class CorrectCompanyGenerator implements GreeterGeneratorInterface
 {
     private CustomNameGenerator $nameGenerator;
 
+    /**
+     * @param CustomNameGenerator $nameGenerator This will be automagically resolved.
+     *                                           See gacela.php in this Integration test.
+     */
     public function __construct(CustomNameGenerator $nameGenerator)
     {
         $this->nameGenerator = $nameGenerator;
