+++ tests/Integration/Framework/BindingInterfacesWithInnerDependencies/LocalConfig/Domain/Greeter/CustomNameGenerator.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GacelaTest\Integration\Framework\BindingInterfacesWithInnerDependencies\LocalConfig\Infrastructure;
+namespace GacelaTest\Integration\Framework\BindingInterfacesWithInnerDependencies\LocalConfig\Domain\Greeter;
 
 final class CustomNameGenerator
 {
