--- tests/Integration/Framework/BindingInterfacesWithInnerDependencies/LocalConfig/Domain/NumberService.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GacelaTest\Integration\Framework\BindingInterfacesWithInnerDependencies\LocalConfig\Domain;
-
-final class NumberService
-{
-    private GreeterGeneratorInterface $numberGenerator;
-
-    public function __construct(GreeterGeneratorInterface $numberGenerator)
-    {
-        $this->numberGenerator = $numberGenerator;
-    }
-
-    public function generateCompanyAndName(): string
-    {
-        return $this->numberGenerator->company('Gacela');
-    }
-}
