+++ tests/Integration/Framework/BindingInterfacesWithInnerDependencies/gacela.php
@@ -3,9 +3,9 @@
 declare(strict_types=1);
 
 use Gacela\Framework\AbstractConfigGacela;
+use GacelaTest\Integration\Framework\BindingInterfacesWithInnerDependencies\LocalConfig\Domain\Greeter\CorrectCompanyGenerator;
+use GacelaTest\Integration\Framework\BindingInterfacesWithInnerDependencies\LocalConfig\Domain\Greeter\IncorrectCompanyGenerator;
 use GacelaTest\Integration\Framework\BindingInterfacesWithInnerDependencies\LocalConfig\Domain\GreeterGeneratorInterface;
-use GacelaTest\Integration\Framework\BindingInterfacesWithInnerDependencies\LocalConfig\Infrastructure\CorrectCompanyGenerator;
-use GacelaTest\Integration\Framework\BindingInterfacesWithInnerDependencies\LocalConfig\Infrastructure\IncorrectCompanyGenerator;
 
 /**
  * This integration-test does two things:
